Add priors table collated same way as posteriors table.
I am not sure why originally I tried collating the priors
from the database instead of from the data in the run
folders. Using the run folders is preferable in that it
only includes the PFTs that were actually used in the run
and not the extraneous pfts that are in BETYdb.